requirements.txt
# Core dependencies
pyspark==3.5.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Configuration
pyyaml==6.0.1

# Data download
kagglehub==0.4.2

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1

# Additional analytics
scipy==1.11.4
scikit-learn==1.3.2

# Utilities
python-dateutil==2.8.2
pyarrow==14.0.1